Remove CoderabbitInstructionsRule; route .coderabbit.yaml through content-* rules
Instead of duplicating weak-language patterns in a separate
CoderabbitInstructionsRule, .coderabbit.yaml instruction text now flows
through the shared content analysis infrastructure:

- _extract_instructions() and helpers moved from coderabbit.py to
  content_analysis.py
- _get_body() handles .coderabbit.yaml by extracting and concatenating
  all instruction field values
- gather_all_content_files() includes .coderabbit.yaml as a "coderabbit"
  category
- All content-* rules (weak-language, tautological, etc.) now
  automatically apply to CodeRabbit instruction text
- Removed coderabbit-instructions from default config and BUILTIN_RULES
- Updated tests and regenerated docs
