+++ .skillsaw.yaml.example
@@ -214,11 +214,6 @@ rules:
     enabled: auto
     severity: error
 
-  # Check .coderabbit.yaml instruction fields for content quality issues
-  coderabbit-instructions:
-    enabled: auto
-    severity: warning
-
 # Load custom rules from these files
 custom-rules: []
